Add tower removing + various other improvements
+ Updated to 1.19.2
+ Add a favicon to the test run coz why not
~ Update MongoDB driver

+ Add tower removing/destroying
~ Change tower IDs to integers as running out of shorts is more likely from someone spamming removing/adding towers (plz no)
~ Changed damage indicators back to just work with Minestom's velocity system

+++ src/main/java/pink/zak/minestom/towerdefence/model/tower/placed/PlacedAttackingTower.java
@@ -22,7 +22,7 @@ public abstract class PlacedAttackingTower<T extends AttackingTowerLevel> extend
     protected List<LivingEnemyMob> targets = new ArrayList<>();
     protected Task attackTask;
 
-    protected PlacedAttackingTower(Instance instance, AttackingTower tower, Material towerPlaceMaterial, short id, GameUser owner, Point baseBlock, Direction facing, int level) {
+    protected PlacedAttackingTower(Instance instance, AttackingTower tower, Material towerPlaceMaterial, int id, GameUser owner, Point baseBlock, Direction facing, int level) {
         super(instance, tower, towerPlaceMaterial, id, owner, baseBlock, facing, level);
         this.startFiring();
     }
