+++ src/main/java/pink/zak/minestom/towerdefence/model/tower/placed/types/BomberTower.java
@@ -27,7 +27,7 @@ public class BomberTower extends PlacedAttackingTower<AttackingTowerLevel> {
     private final MobHandler mobHandler;
     private final GameUser owner;
 
-    public BomberTower(GameHandler gameHandler, Instance instance, AttackingTower tower, Material towerPlaceMaterial, short id, GameUser owner, Point baseBlock, Direction facing, int level) {
+    public BomberTower(GameHandler gameHandler, Instance instance, AttackingTower tower, Material towerPlaceMaterial, int id, GameUser owner, Point baseBlock, Direction facing, int level) {
         super(instance, tower, towerPlaceMaterial, id, owner, baseBlock, facing, level);
         this.mobHandler = gameHandler.getMobHandler();
         this.owner = owner;
